# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SQLoptimize API",
    "description": "SQLoptimize API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/schemas/active": {
      "get": {
        "tags": [
          "schemas"
        ],
        "summary": "获取用户当前活跃的数据库模式",
        "description": "获取用户当前活跃的数据库模式",
        "operationId": "get_active_db_schema_api_schemas_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveDbSchemaResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "schemas"
        ],
        "summary": "设置用户当前活跃的数据库模式",
        "description": "设置用户当前活跃的数据库模式",
        "operationId": "set_active_db_schema_api_schemas_active_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetActiveDbSchemaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetActiveDbSchemaResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/schemas/": {
      "get": {
        "tags": [
          "schemas"
        ],
        "summary": "获取用户数据库模式列表",
        "description": "获取当前用户的数据库模式列表",
        "operationId": "get_user_schemas_api_schemas__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbSchemaListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "schemas"
        ],
        "summary": "创建数据库模式",
        "description": "创建新的数据库模式",
        "operationId": "create_db_schema_api_schemas__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DbSchemaCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbSchemaResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/schemas/{schema_id}": {
      "get": {
        "tags": [
          "schemas"
        ],
        "summary": "根据ID获取数据库模式",
        "description": "获取指定的数据库模式信息",
        "operationId": "get_db_schema_api_schemas__schema_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "schema_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Schema Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbSchemaResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "schemas"
        ],
        "summary": "更新数据库模式",
        "description": "更新数据库模式信息",
        "operationId": "update_db_schema_api_schemas__schema_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "schema_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Schema Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DbSchemaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbSchemaResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "schemas"
        ],
        "summary": "删除数据库模式",
        "description": "删除数据库模式",
        "operationId": "delete_db_schema_api_schemas__schema_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "schema_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Schema Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbSchemaDeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActiveDbSchemaResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "模式ID"
          },
          "schema_name": {
            "type": "string",
            "title": "Schema Name",
            "description": "模式名称"
          },
          "schema_content": {
            "type": "string",
            "title": "Schema Content",
            "description": "模式内容"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "更新时间"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "用户ID"
          }
        },
        "type": "object",
        "required": [
          "id",
          "schema_name",
          "schema_content",
          "created_at",
          "updated_at",
          "user_id"
        ],
        "title": "ActiveDbSchemaResponse",
        "description": "活跃数据库模式响应"
      },
...
      "DbSchemaCreate": {
        "properties": {
          "schema_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Name",
            "description": "模式名称，不提供则自动生成UUID"
          },
          "schema_content": {
            "type": "string",
            "title": "Schema Content",
            "description": "模式内容（JSON字符串）"
          }
        },
        "type": "object",
        "required": [
          "schema_content"
        ],
        "title": "DbSchemaCreate",
        "description": "创建数据库模式请求"
      },
      "DbSchemaDeleteResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "删除结果消息"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "DbSchemaDeleteResponse",
        "description": "数据库模式删除响应"
      },
      "DbSchemaListResponse": {
        "properties": {
          "schemas": {
            "items": {
              "$ref": "#/components/schemas/DbSchemaResponse"
            },
            "type": "array",
            "title": "Schemas",
            "description": "模式列表"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "总数"
          },
          "skip": {
            "type": "integer",
            "title": "Skip",
            "description": "跳过数量"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "限制数量"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "是否还有更多数据"
          },
          "active_schema_id": {
            "type": "integer",
            "title": "Active Schema Id",
            "description": "当前用户活跃的数据库模式ID，0表示无活跃模式",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "schemas",
          "total",
          "skip",
          "limit",
          "has_more"
        ],
        "title": "DbSchemaListResponse",
        "description": "数据库模式列表响应"
      },
      "DbSchemaResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "模式ID"
          },
          "schema_name": {
            "type": "string",
            "title": "Schema Name",
            "description": "模式名称"
          },
          "schema_content": {
            "type": "string",
            "title": "Schema Content",
            "description": "模式内容"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "更新时间"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "用户ID"
          }
        },
        "type": "object",
        "required": [
          "id",
          "schema_name",
          "schema_content",
          "created_at",
          "updated_at",
          "user_id"
        ],
        "title": "DbSchemaResponse",
        "description": "数据库模式响应"
      },
      "DbSchemaUpdate": {
        "properties": {
          "schema_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Name",
            "description": "模式名称"
          },
          "schema_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Content",
            "description": "模式内容（JSON字符串）"
          }
        },
        "type": "object",
        "title": "DbSchemaUpdate",
        "description": "更新数据库模式请求"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SetActiveDbSchemaRequest": {
        "properties": {
          "schema_id": {
            "type": "integer",
            "title": "Schema Id",
            "description": "模式ID"
          }
        },
        "type": "object",
        "required": [
          "schema_id"
        ],
        "title": "SetActiveDbSchemaRequest",
        "description": "设置活跃数据库模式请求"
      },
      "SetActiveDbSchemaResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "设置结果消息"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "SetActiveDbSchemaResponse",
        "description": "设置活跃数据库模式响应"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}